Adds support for NETSTANDARD1.5 / NETCOREAPP1.0

+++ Insight.Database.Core/DbDataReaderWrapper.cs
@@ -63,7 +63,7 @@ namespace Insight.Database
 			// finally copy the data
 			length = Math.Min((int)length, array.Length - (int)dataOffset);
 			if (length > 0)
-				Array.Copy(array, dataOffset, buffer, bufferOffset, length);
+				Array.Copy(array, (int)dataOffset, buffer, bufferOffset, length);
 
 			return length;
 		}
